requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
pypdf
python-dotenv
google-generativeai
streamlit
requests